Pico CLI should work now just need a connection to test with
Made a minor change to validate statements.

Decided tests written already for this class are not in scope, class just relies on data being passed to it by the DB and was testing that. Just needs to be tested for invalid and illegal statement entry blocking.

changed tests written to deal with DB testing in general

#story[817]